requirements.txt
Django==1.7.1
django-extensions==1.4.6
django-debug-toolbar==1.2.2
psycopg2==2.5.4
ipython==2.3.1
Pillow==2.6.1
django-logentry-admin==0.1.4
django-tinymce==1.5.3
Werkzeug==0.9.6
gunicorn==19.1.1
